Convert tile entities and items to fully use Long

+++ src/main/java/moze_intel/projecte/api/capabilities/IKnowledgeProvider.java
@@ -64,12 +64,12 @@ public interface IKnowledgeProvider extends INBTSerializable<NBTTagCompound>
     /**
      * @return The emc in this player's transmutation tablet network
      */
-    double getEmc();
+    long getEmc();
 
     /**
      * @param emc The emc to set in this player's transmutation tablet network
      */
-    void setEmc(double emc);
+    void setEmc(long emc);
 
     /**
      * @param player The player to sync to.
