+++ src/main/java/moze_intel/projecte/api/item/IItemEmc.java
@@ -17,7 +17,7 @@ public interface IItemEmc
 	 * @param toAdd The maximum amount to add
 	 * @return The amount that was actually added
 	 */
-	double addEmc(@Nonnull ItemStack stack, double toAdd);
+	long addEmc(@Nonnull ItemStack stack, long toAdd);
 
 	/**
 	 * Extracts EMC from the itemstack
@@ -25,19 +25,19 @@ public interface IItemEmc
 	 * @param toRemove The maximum amount to remove
 	 * @return The amount that was actually extracted
 	 */
-	double extractEmc(@Nonnull ItemStack stack, double toRemove);
+	long extractEmc(@Nonnull ItemStack stack, long toRemove);
 
 	/**
 	 * Gets the current EMC this stack is showing to the public
 	 * @param stack The stack to query
 	 * @return The current publicly-accessible EMC stored in this stack
 	 */
-	double getStoredEmc(@Nonnull ItemStack stack);
+	long getStoredEmc(@Nonnull ItemStack stack);
 
 	/**
 	 * Gets the maximum EMC that is allowed to be stored in this stack
 	 * @param stack The stack to query
 	 * @return The maximum amount of publicly-accessible EMC that can be stored in this stack
 	 */
-	double getMaximumEmc(@Nonnull ItemStack stack);
+	long getMaximumEmc(@Nonnull ItemStack stack);
 }
