+++ src/main/java/moze_intel/projecte/api/tile/TileEmcHandler.java
@@ -13,41 +13,41 @@ public class TileEmcHandler extends TileEmcBase implements IEmcAcceptor, IEmcPro
 {
 	public TileEmcHandler()
 	{
-		this.maximumEMC = Double.MAX_VALUE;
+		this.maximumEMC = Long.MAX_VALUE;
 	}
 
-	public TileEmcHandler(double max)
+	public TileEmcHandler(long max)
 	{
 		this.maximumEMC = max;
 	}
 
 	// -- IEMCAcceptor -- //
 	@Override
-	public double acceptEMC(@Nonnull EnumFacing side, double toAccept)
+	public long acceptEMC(@Nonnull EnumFacing side, long toAccept)
 	{
-		double toAdd = Math.min(maximumEMC - currentEMC, toAccept);
+		long toAdd = Math.min(maximumEMC - currentEMC, toAccept);
 		currentEMC += toAdd;
 		return toAdd;
 	}
 
 	// -- IEMCProvider -- //
 	@Override
-	public double provideEMC(@Nonnull EnumFacing side, double toExtract)
+	public long provideEMC(@Nonnull EnumFacing side, long toExtract)
 	{
-		double toRemove = Math.min(currentEMC, toExtract);
+		long toRemove = Math.min(currentEMC, toExtract);
 		currentEMC -= toRemove;
 		return toRemove;
 	}
 
 	// -- IEMCStorage --//
 	@Override
-	public double getStoredEmc()
+	public long getStoredEmc()
 	{
 		return currentEMC;
 	}
 
 	@Override
-	public double getMaximumEmc()
+	public long getMaximumEmc()
 	{
 		return maximumEMC;
 	}
