+++ src/main/java/moze_intel/projecte/gameObjs/container/slots/transmutation/SlotConsume.java
@@ -27,8 +27,8 @@ public class SlotConsume extends SlotItemHandler
 		}
 		
 		ItemStack cache = stack.copy();
-		
-		double toAdd = 0;
+
+		long toAdd = 0;
 		
 		while (!inv.hasMaxedEmc() && stack.getCount() > 0)
 		{
