+++ src/main/java/moze_intel/projecte/gameObjs/container/slots/transmutation/SlotInput.java
@@ -38,7 +38,7 @@ public class SlotInput extends SlotItemHandler
 		if (stack.getItem() instanceof IItemEmc)
 		{
 			IItemEmc itemEmc = ((IItemEmc) stack.getItem());
-			double remainingEmc = itemEmc.getMaximumEmc(stack) - (int) Math.ceil(itemEmc.getStoredEmc(stack));
+			long remainingEmc = itemEmc.getMaximumEmc(stack) - (int) Math.ceil(itemEmc.getStoredEmc(stack));
 			
 			if (inv.provider.getEmc() >= remainingEmc)
 			{
