+++ src/main/java/moze_intel/projecte/gameObjs/tiles/RelayMK1Tile.java
@@ -135,7 +135,7 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 			if(stack.getItem() instanceof IItemEmc)
 			{
 				IItemEmc itemEmc = ((IItemEmc) stack.getItem());
-				double emcVal = itemEmc.getStoredEmc(stack);
+				long emcVal = itemEmc.getStoredEmc(stack);
 				
 				if (emcVal > chargeRate)
 				{
@@ -185,9 +185,9 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 	private void chargeItem(ItemStack chargeable)
 	{
 		IItemEmc itemEmc = ((IItemEmc) chargeable.getItem());
-		double starEmc = itemEmc.getStoredEmc(chargeable);
-		double maxStarEmc = itemEmc.getMaximumEmc(chargeable);
-		double toSend = this.getStoredEmc() < chargeRate ? this.getStoredEmc() : chargeRate;
+		long starEmc = itemEmc.getStoredEmc(chargeable);
+		long maxStarEmc = itemEmc.getMaximumEmc(chargeable);
+		long toSend = this.getStoredEmc() < chargeRate ? this.getStoredEmc() : chargeRate;
 			
 		if ((starEmc + toSend) <= maxStarEmc)
 		{
@@ -202,7 +202,7 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 		}
 	}
 
-	public double getItemChargeProportion()
+	public long getItemChargeProportion()
 	{
 		if (!getCharging().isEmpty() && getCharging().getItem() instanceof IItemEmc)
 		{
@@ -212,7 +212,7 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 		return 0;
 	}
 
-	public double getInputBurnProportion()
+	public long getInputBurnProportion()
 	{
 		if (getBurn().isEmpty())
 		{
@@ -224,7 +224,7 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 			return ((IItemEmc) getBurn().getItem()).getStoredEmc(getBurn()) / ((IItemEmc) getBurn().getItem()).getMaximumEmc(getBurn());
 		}
 
-		return getBurn().getCount() / (double) getBurn().getMaxStackSize();
+		return getBurn().getCount() / getBurn().getMaxStackSize();
 	}
 	
 	@Override
@@ -246,7 +246,7 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 	}
 
 	@Override
-	public double acceptEMC(@Nonnull EnumFacing side, double toAccept)
+	public long acceptEMC(@Nonnull EnumFacing side, long toAccept)
 	{
 		if (world.getTileEntity(pos.offset(side)) instanceof RelayMK1Tile)
 		{
@@ -254,16 +254,16 @@ public class RelayMK1Tile extends TileEmc implements IEmcAcceptor, IEmcProvider
 		}
 		else
 		{
-			double toAdd = Math.min(maximumEMC - currentEMC, toAccept);
+			long toAdd = Math.min(maximumEMC - currentEMC, toAccept);
 			currentEMC += toAdd;
 			return toAdd;
 		}
 	}
 
 	@Override
-	public double provideEMC(@Nonnull EnumFacing side, double toExtract)
+	public long provideEMC(@Nonnull EnumFacing side, long toExtract)
 	{
-		double toRemove = Math.min(currentEMC, toExtract);
+		long toRemove = Math.min(currentEMC, toExtract);
 		currentEMC -= toRemove;
 		return toRemove;
 	}
